LayoutTests/platform/ios-simulator/fast/dom/inline-event-attributes-release-expected.txt

Tests that we do not hold on to any nodes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL afterCount - beforeCount should be 0. Was 3.
PASS successfullyParsed is true

TEST COMPLETE
